client/requirements.txt
pyserial
python-dotenv
pytz
env-canada
feedparser